search-provider: Initial implementation
The implementation is largely based on gnome-weather.  Unlike other
search-providers, this initiates a new search upon both
GetInitialResultSet and GetSubsearchResultSet, because the initial
result set can be too large and refinement doesn't work.

Also, upon ActivateResult it doesn't launch the application, but copies
the character into clipboard, as suggested in bug 744855.