SearchProvider: use proper startup notification for launching search
This code was copied from gnome-weather, which explicitly tries to
avoid Gdk/Gtk+ in the background service (so that it could be
one day launched as a user service from systemd). But gnome-characters
initializes Gtk+ anyway to handle the clipboard, there is no
reason for degraded operation.
In particular, this gives a nice spinner while gnome-characters
starts up (which is almost invisible, but it can become visible
on a slow disk and cold cache)

https://bugzilla.gnome.org/show_bug.cgi?id=745508