characterList: fix initial load length in big windows
When the window is big (eg. maximized at 1080p), the character list
initially loads only 100 items, when the screen fits at least 200.
This is a bad experience, as it makes the user think there are only a
small amount of characters in that category.  Additionally, it breaks
the lazy-loading; as _onEdgeReached is never called because the user can
not scroll.

To fix this, we need to calculate the amount of characters a screen
fits, and use that number instead.

https://bugzilla.gnome.org/show_bug.cgi?id=785879