libgc: Perform full canonical decomposition for Hangul syllables
Previously, the code finding related characters only took into account
of composed characters built from a base character and combining
characters (such as Latin, Hiragana, and Katakana).  However, Hangul
syllables are composed of two or three Hangul jamo characters, all of
which should be considered as a base character.  This patch handles
that case properly.

For the implementation, uc_canonical_decomposition() is not capable of
decomposing Hangul syllables.  Instead of the function, this patch
uses u32_normalize() with UNINORM_NFD, as suggested by Bruno Haible in:
https://lists.gnu.org/archive/html/bug-libunistring/2017-11/msg00002.html

https://bugzilla.gnome.org/show_bug.cgi?id=790391