build: Reuse characters name variable
gnome-characters uses an input file as a template to generate a
service file that has two tags for `PACKAGE_NAME` and `pkgdatadir`.

The `PACKAGE_NAME` tag should be replace with the project name, so
meson's build file uses the `project_name` function to access it.
However, there is already a variable to access faster to this value
and to maintain it consistently among every use of it.

This has been changed so `PACKAGE_NAME` also uses that variable.