build: Rename project name
gnome-characters has been using its namespace as its name which eases
some building steps in autotools. This might cause some issues with
meson because it uses this value in a different way for some steps.

However, meson is able to separate these concepts so the project name
can been changed and use the namespace in different build steps.

The project name has been renamed to `gnome-characters` and the
namespace is used for the necessary build steps.