searchProvider: Keep the application active while searching
Prevent the application to be closed while the search provider is
performing any action. So that the search is always completed.

This should also prevent that we're destroying the Gsk.GLRenderer before
we've correctly unrealized it.

To avoid the releasing to stop the application immediately, we also need
to explicitly set the inactivity timeout (using GApplication default).

Helps with: https://gitlab.gnome.org/GNOME/gnome-characters/-/issues/124