lib: inline parsing of hex code-point
This gets called rougly 290,000 times on a single search for one keyword
and therefore could use some optimization. This does both the character
check and the parsing in the same pass. Normally, that might be more
expensive because you have to do extra work than just checking characters,
but strspn() returns the offset in the substring and therefore is doing
additional unnecessary work anyway and this will likely bail on the first
character in the negative case.