lib: reduce calls to g_task_return_error_if_cancelled()
For a simple search result in the gnome-characters UI, we may easily have
20,000 search results (with max-results set to G_MAXUINT). This results
in a huge amount of calls to g_task_return_error_if_cancelled() which can
be as expensive as the other work being done.

Instead, check it infrequently to get the non-local function call out of
the hot path. 2048 was chosen because things are fast enough now that we
can get a bunch of work done in that time without losing much latency.
Additionally, 2048 will turn into `iterations & 0x7FF == 0` when compiled.