lib: avoid utf-8 conversion and allocation in hot path
This code checks to see if the keyword is an exact match to the ucs4 value
converted to utf-8. Previously, that code would use g_ucs4_to_utf8() to do
that conversion which gives us the length of the string and a malloc()'d
version of the UTF-8 form.

Instead of allocating the UTF-8 version of the target we can instead
directly compare each ucs4 value to the unichar format of our keyword.

There is a further optimization to be had by pre-calculating the ucs4
version of all the keywords so we can compare them without having to
extract the unichar from the UTF-8 string each time. But that will require
a bit more abstraction on the criteria and my changes to do that weren't
very pretty. So if that is wanted, I'll leave that up to someone who really
digs that sort of work.