searchProvider: reduce MAX_SEARCH_RESULTS to 20
This type of search is highly interactive. The shell only shows 5 matches
for a search provider and so this number only needs to be high enough to
show the user that there are more entries within the application that they
can take a look at. 15 is as good as 95 from this standpoint and it allows
us to bail earlier to process another incoming request.