Add formatting and linting settings
While working on an unrelated change, I was fighting with my editor
constantly wanting to reformat the code. So as a prerequisite, here
we're adding an uncrustify configuration file (taken from Mutter) and a
`.editorconfig` file. The generation scripts have also been updated to
generate code which is compliant with this new formatting, and output
the headers that are required, so each header can be opened and IDEs
can resolve symbols.

A CI job to lint these file has been added, and the scripts to generate
the headers updated to produce files that are compatible.