chore: add formatting and linting settings for Python files
As with the previous commit, editors are trying to lint/reformat Python
files. To have a clean base to work from, here we add a
`pyproject.toml` file to configure Ruff, a CI job to enforce
cleanliness, and then fix all of the current findings to have a clean
baseline.