chore(js): update to ESM, modern GJS and new ESLint
GJS now supports ESM, replacing the nonstandard `imports` ayntax for
importing dependencies. This means we can do:

```javascript
import { Gio, GLib, GObject } from 'gi://Gio?version=2.0';

export
```

instead of

```javascript
const { Gio, GLib, GObject } = imports.gi;
```

here we port to that.

Additionally, ESLint has had a major upgrade with a new config system.
This commit takes the config from `gnome-shell`, slims it down a bit to
our needs, and sets it up to be in use by IDEs and in CI. Then it was
run with `--fix` to automatically fix most of the issues, with the rest
manually cleaned up.